Minor fixes
Added heap leaks detection to testing app.
Added missing destructor into Spline Decompressor.
Added some warning blokes.
Fixed chunk owerflow not being detected for new formats.

+++ source/hkNewHeader.cpp
@@ -195,7 +195,7 @@ int chTNAMRead(BinReader * rd, hkChunk *holder, hkxNewHeader *root)
 		}		
 	}
 
-	const size_t diff = holder->Size() - (rd->Tell() - savepos);
+	const int diff = static_cast<int>(holder->Size()) - static_cast<int>(rd->Tell() - savepos);
 
 	if (diff < 0)
 	{
